requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
filelock==3.9.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
psycopg2==2.9.5
PyJWT==2.6.0
python-decouple==3.7
pytils==0.4.1
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.19.0
